Make model and rename tests
