+++ QuickCalUITests/ContentViewUITests.swift
@@ -1,0 +1,23 @@
+//
+//  ContentViewUITests.swift
+//  ContentViewUITests
+//
+//  Created by Will Boland on 9/11/22.
+//
+
+import XCTest
+
+class ContentViewUITests: XCTestCase {
+
+    override func setUpWithError() throws {
+        continueAfterFailure = false
+    }
+
+    func testContentView_Success() throws {
+        let app = XCUIApplication()
+        app.launch()
+
+        let welcomeText = app.staticTexts["Welcome!"]
+        XCTAssert(welcomeText.exists)
+    }
+}
