--- QuickCalUITests/QuickCalUITests.swift
@@ -1,44 +1,0 @@
-//
-//  QuickCalUITests.swift
-//  QuickCalUITests
-//
-//  Created by Will Boland on 9/11/22.
-//
-
-import XCTest
-
-class QuickCalUITests: XCTestCase {
-
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-
-        // In UI tests it is usually best to stop immediately when a failure occurs.
-        continueAfterFailure = false
-
-        // In UI tests you should set the initial state (orientation, etc) required for your tests before they run.
-        // The setUp method is a good place to do this.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testExample() throws {
-        // UI tests must launch the application that they test.
-        let app = XCUIApplication()
-        app.launch()
-
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-        let welcomeText = app.staticTexts["Welcome!"]
-        XCTAssert(welcomeText.exists)
-    }
-
-    func testLaunchPerformance() throws {
-        if #available(macOS 10.15, iOS 13.0, tvOS 13.0, watchOS 7.0, *) {
-            // This measures how long it takes to launch your application.
-            measure(metrics: [XCTApplicationLaunchMetric()]) {
-                XCUIApplication().launch()
-            }
-        }
-    }
-}
