--- QuickCalUITests/QuickCalUITestsLaunchTests.swift
@@ -1,32 +1,0 @@
-//
-//  QuickCalUITestsLaunchTests.swift
-//  QuickCalUITests
-//
-//  Created by Will Boland on 9/11/22.
-//
-
-import XCTest
-
-class QuickCalUITestsLaunchTests: XCTestCase {
-
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        true
-    }
-
-    override func setUpWithError() throws {
-        continueAfterFailure = false
-    }
-
-    func testLaunch() throws {
-        let app = XCUIApplication()
-        app.launch()
-
-        // Insert steps here to perform after app launch but before taking a screenshot,
-        // such as logging into a test account or navigating somewhere in the app
-
-        let attachment = XCTAttachment(screenshot: app.screenshot())
-        attachment.name = "Launch Screen"
-        attachment.lifetime = .keepAlways
-        add(attachment)
-    }
-}
